backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
google-cloud-datastore==2.21.0
python-dotenv==1.0.1
pydantic==2.9.2
python-chess==1.999
torch
numpy